Proxying HTTPS requests over secure TCP tunnel (HTTP CONNECT)
This is implemented by using https-proxy-agent as a tunneling agent.

When an https.Agent is passed in the Axios config, its options are
passed along to the HttpsProxyAgent instance to preserve options like
ca and rejectUnauthorized (kind of a hack, explicit options would
be better).

Out of scope of this work: adding an option to force using a
tunnel for (non-secure) HTTP requests, or explicitly disable
tunnelling for (secure) HTTPS requests.